added acos() check to ticket #1513; added testcases for MathLib::toDoubleNumber(), now double conversion of zeros is handled correctly; changed assertEquals() function of testsuite parameters from unsigned int to double. This is needed to avoid overflow of unsigned int by comparing negative floating point values.

+++ test/testsuite.cpp
@@ -118,7 +118,7 @@ void TestFixture::assertEquals(const char *filename, int linenr, const std::stri
     }
 }
 
-void TestFixture::assertEquals(const char *filename, int linenr, unsigned int expected, unsigned int actual)
+void TestFixture::assertEquals(const char *filename, int linenr, double expected, double actual)
 {
     std::ostringstream ostr1;
     ostr1 << expected;
